moved service implementation to seperate file
